Revert "Revert renames in client"
This reverts commit cc0a95d14c3212fa1c49f93789dd94e167427a57.
