+++ .gitattributes
@@ -1,1 +1,1 @@
-client/lcd/swagger-ui/* linguist-vendored
+client/docs/swagger-ui/* linguist-vendored
