+++ client/docs/statik/statik.go
