+++ client/docs/swagger-ui/favicon-16x16.png
