+++ client/docs/swagger-ui/favicon-32x32.png
