+++ client/docs/swagger-ui/oauth2-redirect.html
