--- client/lcd/statik/init.go
@@ -1,3 +1,0 @@
-package statik
-
-//This just for fixing the error in importing empty github.com/cosmos/cosmos-sdk/client/lcd/statik
